Split up bulky files under src/media into .h/.cpp files.
