+++ src/media/MidiDisplayComponent.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <juce_audio_basics/juce_audio_basics.h>
+
 #include "../pianoroll/PianoRollComponent.hpp"
 #include "MediaDisplayComponent.h"
 
@@ -8,175 +10,38 @@ class MidiDisplayComponent : public MediaDisplayComponent
 {
 public:
 
-    MidiDisplayComponent()
-    {
-        addAndMakeVisible(pianoRoll);
+    MidiDisplayComponent();
+    ~MidiDisplayComponent();
 
-        mediaHandlerInstructions = "MIDI pianoroll.\nClick and drag to start playback from any point in the pianoroll\nVertical scroll to zoom in/out.\nHorizontal scroll to move the pianoroll.";
-    }
+    static StringArray getSupportedExtensions();
+    StringArray getInstanceExtensions() { return MidiDisplayComponent::getSupportedExtensions(); }
 
-    ~MidiDisplayComponent() {}
+    void drawMainArea(Graphics& g, Rectangle<int>& a) override;
+    void resized() override;
 
-    void drawMainArea(Graphics& g, Rectangle<int>& a) override
-    {
-        pianoRoll.setBounds(a);
-    }
+    void loadMediaFile(const URL& filePath) override;
 
-    void resized() override
-    {
-        Rectangle<int> scrollBarArea = getLocalBounds().removeFromBottom(scrollBarSize + 2 * scrollBarSpacing);
-        scrollBarArea = scrollBarArea.removeFromRight(scrollBarArea.getWidth() - pianoRoll.getKeyboardWidth() - 5);
-        scrollBarArea = scrollBarArea.removeFromLeft(scrollBarArea.getWidth() - 2 * pianoRoll.getScrollBarSize() - 4 * pianoRoll.getScrollBarSpacing());
+    void setPlaybackPosition(double t) override;
+    double getPlaybackPosition() override;
 
-        horizontalScrollBar.setBounds(scrollBarArea.reduced(scrollBarSpacing));
-    }
+    bool isPlaying() override;
+    void startPlaying() override;
+    void stopPlaying() override;
 
-    static StringArray getSupportedExtensions()
-    {
-        StringArray extensions;
+    double getTotalLengthInSecs() override { return totalLengthInSecs; }
 
-        extensions.add(".mid");
-        extensions.add(".midi");
-
-        return extensions;
-    }
-
-    StringArray getInstanceExtensions()
-    {
-        return MidiDisplayComponent::getSupportedExtensions();
-    }
-
-    void loadMediaFile(const URL& filePath) override
-    {
-        // Create the local file this URL points to
-        File file = filePath.getLocalFile();
-
-        std::unique_ptr<juce::FileInputStream> fileStream(file.createInputStream());
-
-        // Read the MIDI file from the File object
-        MidiFile midiFile;
-
-        if (!midiFile.readFrom(*fileStream)) {
-            DBG("Failed to read MIDI data from file.");
-        }
-
-        midiFile.convertTimestampTicksToSeconds();
-
-        totalLengthInSecs = midiFile.getLastTimestamp();
-
-        DBG("Total duration of MIDI file " << totalLengthInSecs << " seconds.");
-
-        pianoRoll.resizeNoteGrid(totalLengthInSecs);
-
-        for (int trackIdx = 0; trackIdx < midiFile.getNumTracks(); ++trackIdx) {
-            const juce::MidiMessageSequence* constTrack = midiFile.getTrack(trackIdx);
-
-            if (constTrack != nullptr) {
-                juce::MidiMessageSequence track(*constTrack);
-                track.updateMatchedPairs();
-
-                DBG("Track " << trackIdx << " has " << track.getNumEvents() << " events.");
-
-                for (int eventIdx = 0; eventIdx < track.getNumEvents(); ++eventIdx) {
-                    const auto midiEvent = track.getEventPointer(eventIdx);
-                    const auto& midiMessage = midiEvent->message;
-
-                    double startTime = midiEvent->message.getTimeStamp();
-
-                    DBG("Event " << eventIdx << " at " << startTime << ": " << midiMessage.getDescription());
-
-                    if (midiMessage.isNoteOn()) {
-                        int noteNumber = midiMessage.getNoteNumber();
-                        int velocity = midiMessage.getVelocity();
-
-                        double duration = 0;
-
-                        for (int offIdx = eventIdx + 1; offIdx < track.getNumEvents(); ++offIdx) {
-                            const auto offEvent = track.getEventPointer(offIdx);
-
-                            // Find the matching note off event
-                            if (offEvent->message.isNoteOff() && offEvent->message.getNoteNumber() == noteNumber) {
-                                duration = (offEvent->message.getTimeStamp() - midiEvent->message.getTimeStamp());
-                                break;
-                            }
-                        }
-
-                        // Create a component for each for each note
-                        MidiNoteComponent n = MidiNoteComponent(noteNumber, velocity, startTime, duration);
-                        pianoRoll.insertNote(n);
-                    }
-                }
-            }
-        }
-    }
-
-    void setPlaybackPosition(double t) override
-    {
-        // TODO
-    }
-
-    double getPlaybackPosition() override
-    {
-        // TODO
-        return 0.0;
-    }
-
-    void startPlaying() override
-    {
-        // TODO
-        AlertWindow::showMessageBoxAsync(
-            AlertWindow::WarningIcon,
-            "NotImplementedError",
-            "MIDI playback has not yet been implemented."
-        );
-    }
-
-    void stopPlaying() override
-    {
-        // TODO
-    }
-
-    bool isPlaying() override
-    {
-        // TODO
-        return false;
-    }
-
-    double getTotalLengthInSecs() override
-    {
-        return totalLengthInSecs;
-    }
-
-    void updateVisibleRange(Range<double> newRange) override
-    {
-        MediaDisplayComponent::updateVisibleRange(newRange);
-
-        pianoRoll.updateVisibleMediaRange(newRange);
-    }
+    void updateVisibleRange(Range<double> newRange) override;
 
 private:
 
-    double xToTime(const float x) const override
-    {
-        // TODO
-        return 0.0;
-    }
+    double xToTime(const float x) const override;
+    float timeToX(const double t) const override;
 
-    float timeToX(const double t) const override
-    {
-        // TODO
-        return 0.0f;
-    }
+    void resetDisplay() override;
 
-    void resetDisplay() override
-    {
-        pianoRoll.resetNotes();
-        pianoRoll.resizeNoteGrid(0.0);
-    }
-
-    void postLoadActions(const URL& filePath) override {}
-
-    PianoRollComponent pianoRoll{70, scrollBarSize, scrollBarSpacing};
+    void postLoadActions(const URL& filePath) override;
 
     double totalLengthInSecs;
+
+    PianoRollComponent pianoRoll{70, scrollBarSize, scrollBarSpacing};
 };
